James Bond learns how to set traps.
Add a new macro, with-trap that throws and exception is any of the
supplied functions are called.

The rational for addign this macro is to avoid this common pattern when
testing Clojure:

    (bond/with-stub [foo]
      (bar)
      (is (zero? (foo bond/calls count))))

This can be replaced with the following:

    (bond/with-trap [foo]
      (bar)

If foo is not called by bar in this example, then the code will behave as normal
and the result of bar will be returned from the bond/with-trap form. If
foo is called, then bond/with-trap will throw an exception.