Useful error message for when stub!/with-stub! will fail
If a function was generated without `:arglists` metadata then it can't
be stubbed using `with-stub!` since `stub!` can't verify the arity of
the replacement function.

Instead of giving an inscrutable error message, we now assert that
`:arglists` exists and produce an error message which points the user in
the right direction.

Fixes: #41