Add local variants spy and stub
`with-stub` and `wih-spy` functions alter the root of the Var of the
function, through with-redefs, which makes this redefinition visible
to all threads.

Now we provide local versions of with-stub and with-spy which function
the same way as the normal versions but the changes are only visible
to the currently running thread. Other threads see the original
redefinitons or their own redefinitions.

This only works for threads which respect the Clojure per-thread
bindings, and thus won't work for other concurrency primitives like
java.lang.Thread.