Remove support for ClojureScript
We no longer actively use ClojureScript within CircleCI and supporting
it in this project has become a non-trivial amount of effort. In
particular, the behaviour of `ns-publics` is different between Clojure
and ClojureScript, after upgrading both:

- https://github.com/clojure/clojure/blob/clojure-1.10.2/src/clj/clojure/core.clj#L4177-L4186
- https://github.com/clojure/clojurescript/blob/r1.10.758/src/main/clojure/cljs/core.cljc#L2903-L2917

Which causes this exception that can't be easily satisfied for both:

    Caused by: java.lang.AssertionError: Assert failed: Argument to ns-publics must be a quoted symbol

I can only find one ClojureScript project that uses a recent version of
bond. We'll release this change in a new version to indicate the break
in compatibility:

- https://github.com/search?q=language%3Aclojure+clojurescript+circleci%2Fbond&type=Code