ci: Automate publishing tags to clojars
So that we don't need to find credentials and remember how to release
changes manually. Git tags are used to trigger publishing so that we
don't re-publish over existing versions every time we merge to master.
It also means that we're more likely to remember to tag and don't need
to add any logic or credentials for the workflow to create tags.

An alternative would have been to use the build number as a patch
version, but it's unwieldy for both for users and for us to maintain the
versions in `README.md` and `CHANGELOG.md`.

The `CLOJARS_USERNAME` and `CLOJARS_TOKEN` environment variables have
been added to the project. At the moment they belong to an individual
but we'll move them to a machine user later. The project already has
"build forks" and "pass secrets to fork builds" disabled, which will
prevent leaking these credentials.